address update

+++ app/routes/address.php
@@ -1,7 +1,8 @@
 <?php
-//use Model\User;
-//use Controller\UserController;
 
+/* *
+ * Get User Addresses
+ * */
 $app->get('/address[/{uuid}]', function ( $request, $response, $args) {
 
 	//check user if valid
@@ -11,45 +12,60 @@ $app->get('/address[/{uuid}]', function ( $request, $response, $args) {
 		return $response;
 	}
 
-	$selectUserStatement = $this->db->select()->from('addresses')->where('user_uuid','=',$args['uuid']);
-	$stmt = $selectUserStatement->execute(false);
-	$data = $stmt->fetchAll();
-	
-	return $response->withJson(array("status" => true, "data" => $data), 200);
+	try {
+		
+		$selectUserStatement = $this->db->select()->from('addresses')->where('user_uuid','=',$args['uuid']);
+		$stmt = $selectUserStatement->execute(false);
+		$data = $stmt->fetchAll();
+		
+		return $response->withJson(array("status" => true, "data" => $data), 200);		
+		
+	} catch (Exception $e) {
+		
+		return $response->withJson(array("status" => false, "message" => $e->getMessage()), 500);
+		
+	}
 
 });
 
-$app->post('/address[/{uuid}]', function ( $request, $response, $args) {
+/* *
+ * Add Address
+ * */
+$app->post('/address/{user_uuid}', function ( $request, $response, $args) {
 
 	$data = $request->getParsedBody();
-
+	
 	//check for data, return false if empty
 	if ( empty($data) ) {
-		$response->withJson(array("status" => false, "message" => 'Empty form!'), 404);
+		$response->withJson(array("status" => false, "message" => 'Empty Form!'), 404);
 		return $response;
 	}
 
 	//check user if valid
-	$userId = $this->UserUtil->checkUser($args['uuid']);
+	$userId = $this->UserUtil->checkUser($args['user_uuid']);
 	if ( ! $userId ) {
 		$response->withJson(array("status" => false, "message" => "Invalid User"), 404);
 		return $response;
 	}
 
-	$data['user_uuid'] = $args['uuid'];
+	$data['user_id'] = $userId;
 	$data['created'] = date('Y-m-d h:i:s');
 
 	$arrFields = array_keys($data);
 	$arrValues = array_values($data);
-
+	
 	try {
 		// INSERT INTO address ( id , usr , pwd ) VALUES ( ? , ? , ? )
 		$insertStatement = $this->db->insert( $arrFields )
 									->into('addresses')
 									->values($arrValues);
-
 		$insertId = $insertStatement->execute(true);
-		return $response->withJson(array("status" => true, "data" => $insertId), 200);
+		
+		$selectStmt = $this->db->select()->from('addresses')->where('id','=',$insertId);
+		$selectStmt = $selectStmt->execute();
+		$arrResult = $selectStmt->fetch();
+
+		return $response->withJson(array("status" => true, "data" => $arrResult), 200);
 
 	} catch (PDOException $e) {
 
@@ -59,10 +75,13 @@ $app->post('/address[/{uuid}]', function ( $request, $response, $args) {
 
 });
 
-$app->put('/address[/{id}]', function ( $request, $response, $args) {
+/* *
+ * Update Address
+ * */
+$app->put('/address/{user_uuid}/{id}', function ( $request, $response, $args) {
 
 	$data = $request->getParsedBody();
-
+	
 	//check for data, return false if empty
 	if (empty($data)) {
 		//$response->setStatus(500);
@@ -70,19 +89,37 @@ $app->put('/address[/{id}]', function ( $request, $response, $args) {
 	}
 
 	//check user if valid
-	$userId = $this->UserUtil->checkUser($data['user_uuid']);
+	$userId = $this->UserUtil->checkUser($args['user_uuid']);
 	if ( ! $userId ) {
-		$response->withJson(array("status" => false, "message" => "Invalid User"), 404);
-		return $response;
+		return $response->withJson(array("status" => false, "message" => "Invalid User"), 404);
 	}
-
-	// UPDATE users SET pwd = ? WHERE id = ?
-	$updateStatement = $this->db->update( $data )
-								->table('addresses')
-								->where('id', '=', $args['id']);
-
-	$insertId = $updateStatement->execute(true);
-	return $response->withJson(array("status" => true, "data" => $insertId), 200);
+	
+	//check if the user owns the address_id
+	$selectStmt = $this->db->select()->from('addresses')->whereMany(array('id'=>$args['id'], 'user_id'=>$userId ),'=')->execute();
+	$arrAddress = $selectStmt->fetch();
+	if(!$arrAddress) {
+		return $response->withJson(array("status" => false, "message" => "User Address Not Found!"), 404);		
+	}
+	
+	try {
+		
+		// UPDATE users SET pwd = ? WHERE id = ?
+		$updateStatement = $this->db->update( $data )
+									->table('addresses')
+									->where('id', '=', $args['id']);
+		$updateBln = $updateStatement->execute(true);
+		
+		//Get the address details
+		$selectStmt = $this->db->select()->from('addresses')->where('id','=',$args['id'])->execute();
+		$arrResult = $selectStmt->fetch();
+		
+		return $response->withJson(array("status" => true, "data" => $arrResult), 200);
+		
+	} catch (Exception $e) {
+	
+		return $response->withJson(array("status" => false, "message" => $e->getMessage()), 500);
+		
+	}
 
 });
 
@@ -100,6 +137,14 @@ $app->delete('/address/{user_uuid}/{id}', function ( $request, $response, $args)
 	}
 
 	try {
+		
+		//check if the user owns the address_id
+		$selectStmt = $this->db->select()->from('addresses')->whereMany(array('id'=>$args['id'], 'user_id'=>$userId ),'=')->execute();
+		$arrAddress = $selectStmt->fetch();
+		if(!$arrAddress) {
+			return $response->withJson(array("status" => false, "message" => "User Address Not Found!"), 404);
+		}		
+		
 		// DELETE FROM users WHERE id = ?
 		$deleteStatement = $this->db->delete()
 									->from('addresses')
@@ -109,7 +154,7 @@ $app->delete('/address/{user_uuid}/{id}', function ( $request, $response, $args)
 
 		return $response->withJson(array("status" => true, "data" => $affectedRows), 200);
 
-	} catch (PDOException $e) {
+	} catch (Exception $e) {
 
 		return $response->withJson(array("status" => false, "message" => $e->getMessage()), 500);
